requirements.txt
streamlit
opencv-python-headless
zxing-cpp
numpy
Pillow